refactor: move MultiClockTester
